+++ widgets-samples/cc/samples-cc-react-app/src/App.tsx
@@ -10,7 +10,7 @@ import {
   OutdialCall,
 } from '@webex/cc-widgets';
 import {StationLogoutSuccess} from '@webex/plugin-cc';
-import Webex from 'webex/contact-center';
+import Webex from 'webex';
 import {ThemeProvider, IconProvider, Icon, Button, Checkbox, Text, Select, Option} from '@momentum-design/components/dist/react';
 import {PopoverNext} from '@momentum-ui/react-collaboration';
 import './App.scss';
@@ -85,12 +85,9 @@ const handleSaveEnd = (isComplete: boolean) => {
       const urlParams = new URLSearchParams(window.location.hash.replace('#', '?'));
     
       const accessToken = urlParams.get('access_token');
-      const expiresIn = urlParams.get('expires_in') ?? '0';
     
       if (accessToken) {
         window.localStorage.setItem('accessToken', accessToken);
-        // @ts-expect-error: Browser accepts this
-        window.localStorage.setItem('date', new Date().getTime() + parseInt(expiresIn, 10));
         setAccessToken(accessToken);
         // Clear the hash from the URL to remove the token from browser history
         window.history.replaceState(
@@ -101,14 +98,9 @@ const handleSaveEnd = (isComplete: boolean) => {
       }
     }
     else {
-      const storedDate = window.localStorage.getItem('date');
-      if (storedDate && parseInt(storedDate, 10) > new Date().getTime()) {
-        const storedAccessToken = window.localStorage.getItem('accessToken');
-        if (storedAccessToken) {
-          setAccessToken(storedAccessToken);
-        }
-      } else {
-        window.localStorage.removeItem('accessToken');
+      const storedAccessToken = window.localStorage.getItem('accessToken');
+      if (storedAccessToken) {
+        setAccessToken(storedAccessToken);
       }
     }
   }
@@ -158,16 +150,20 @@ const onTaskDeclined = (task,reason) => {
     setShowRejectedPopup(true);
   };
 
-  const onTaskSelected = (task) => {
-    console.log('onTaskSelected invoked for task:', task);
+  const onTaskSelected = ({task, isClicked}) => {
+    console.log('onTaskSelected invoked for task:', task, 'isClicked:', isClicked);
   };
 
-  const onHoldResume = () => {
-    console.log('onHoldResume invoked');
+  const onHoldResume = ({isHeld, task}) => {
+    console.log('onHoldResume invoked', {isHeld, task});
   };
 
-  const onEnd = () => {
-    console.log('onEnd invoked');
+  const onRecordingToggle = ({isRecording, task}) => {
+    console.log('onRecordingToggle invoked', {isRecording, task});
+  };
+
+  const onEnd = ({task}) => {
+    console.log('onEnd invoked', {task});
   };
 
   const onWrapUp = (params) => {
@@ -306,7 +302,9 @@ const onTaskDeclined = (task,reason) => {
 
   // Store accessToken changes in local storage
   useEffect(() => {
-    window.localStorage.setItem('accessToken', accessToken);
+    if(accessToken.trim() !== '') {
+      window.localStorage.setItem('accessToken', accessToken);
+    }
   }, [accessToken]);
 
   useEffect(() => {
@@ -680,7 +678,7 @@ const onTaskDeclined = (task,reason) => {
                         <section className="section-box">
                           <fieldset className="fieldset">
                             <legend className="legend-box">Call Control</legend>
-                            <CallControl onHoldResume={onHoldResume} onEnd={onEnd} onWrapUp={onWrapUp} />
+                            <CallControl onHoldResume={onHoldResume} onEnd={onEnd} onWrapUp={onWrapUp} onRecordingToggle={onRecordingToggle} />
                           </fieldset>
                         </section>
                       </div>
@@ -694,6 +692,7 @@ const onTaskDeclined = (task,reason) => {
                               onHoldResume={onHoldResume}
                               onEnd={onEnd}
                               onWrapUp={onWrapUp}
+                              onRecordingToggle={onRecordingToggle}
                               callControlClassName={'call-control-outer'}
                               callControlConsultClassName={'call-control-consult-outer'}
                             />